Integrate new event system into Application_2
Event system is the core part of my application, so much
that I decided to make new "Application" class.
Because, world modeling now should be on a separate layer,
and event handling have to be on a separate layer too.

Maybe I should make Model class to be EventReceiver, and push it
instead of "LayerWorld".